requirements.txt
streamlit==1.31.0
pillow==10.1.0
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
uuid==1.30